Switch to Partial TrackingStates for Tracking config
