+++ TF_Tooling_Web/src/Connection/MessageReceiver.ts
@@ -5,7 +5,7 @@ import {
     ResponseCallback,
     ServiceStatus,
     ServiceStatusCallback,
-    TrackingState,
+    TrackingStateRequest,
     TrackingStateCallback,
     TrackingStateResponse,
     WebSocketResponse
